hostapd: Change drv_flags from unsigned int to u64
Some flag already using a bit larger than 32, so extend the hostapd
drv_flags type similarly to the earlier wpa_supplicant change to get the
full flag content.

+++ src/ap/hostapd.h
@@ -308,7 +308,7 @@ struct hostapd_iface {
 	struct ap_info *ap_list; /* AP info list head */
 	struct ap_info *ap_hash[STA_HASH_SIZE];
 
-	unsigned int drv_flags;
+	u64 drv_flags;
 
 	/* SMPS modes supported by the driver (WPA_DRIVER_SMPS_MODE_*) */
 	unsigned int smps_modes;
